requirements.txt
Flask
mysql-connector
google.auth